forward c10 registrations to aten
Pull Request resolved: https://github.com/pytorch/pytorch/pull/25969

When an op is registered with the c10 dispatcher but the op is actually one of the ops that still lives on globalATenDispatch(), the c10 dispatcher will forward the registration call to globalATenDispatch().

This will allow backend extensions like XLA to immediately move to the new c10 registration API, while we are moving ops from globalATenDispatch to the c10 dispatcher behind the scenes without breaking their API.

Note that for consistency, this also changes all op registrations in TypeDefault/CPUType/... They now also directly go to the c10 dispatcher, and will be rerouted to globalATenDispatch from there.

ghstack-source-id: 89952417

Differential Revision: [D17301917](https://our.internmc.facebook.com/intern/diff/D17301917/)